requirements.txt
selenium
undetected-chromedriver
webdriver-manager
datetime
regex
discord
discord.py
python-dotenv
seleniumbase
nodriver
bs4
BeautifulSoup4
pyautogui
pillow
pytesseract
opencv-python-headless
brotli
# Optional: for PROFILE_ENCRYPTION_KEY (encrypt user_profiles.json)
cryptography
# Web app
fastapi
uvicorn[standard]
jinja2
